fix: コントラクトのテストファイル名修正
